// web/src/api/generated/openapi.ts
export interface components {
    schemas: {
// ...
        ApiError: {
            /** Code */
            code: string;
            /** Details */
            details?: {
                [key: string]: unknown;
            };
            /** Message */
            message: string;
            /** Recoverable */
            recoverable: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}